move edit new stuff to new file
