+++ packages/cursorless-engine/src/actions/Actions.ts
@@ -6,7 +6,8 @@ import Call from "./Call";
 import Clear from "./Clear";
 import { CutToClipboard } from "./CutToClipboard";
 import Deselect from "./Deselect";
-import { EditNew, EditNewAfter, EditNewBefore } from "./EditNew";
+import { EditNew } from "./EditNew";
+import { EditNewAfter, EditNewBefore } from "./EditNewLineAction";
 import ExecuteCommand from "./ExecuteCommand";
 import { FindInWorkspace } from "./Find";
 import FollowLink from "./FollowLink";
